Partially moved files to proper places for server loading
