+++ client/main.js
@@ -1,6 +1,3 @@
-import { Template } from 'meteor/templating';
-import { ReactiveVar } from 'meteor/reactive-var';
-
-/*global _*/
-
-import './main.html';
+// NOTE: this is the application entry point for clients. all other html,js,css code is loaded from here
+import '/imports/startup/client';
+import '/imports/startup/both';
