+++ client/main.scss
@@ -1,0 +1,23 @@
+// TODO: use a css file structured like this example
+// @import "{}/imports/ui/stylesheets/reset.less";
+//
+// // Global namespace
+// @import "{}/imports/ui/stylesheets/base.less";
+// @import '{}/imports/ui/stylesheets/button.less';
+// @import '{}/imports/ui/stylesheets/form.less';
+// @import '{}/imports/ui/stylesheets/icon.less';
+// @import '{}/imports/ui/stylesheets/layout.less';
+// @import '{}/imports/ui/stylesheets/link.less';
+// @import '{}/imports/ui/stylesheets/menu.less';
+// @import '{}/imports/ui/stylesheets/nav.less';
+//
+// // Global components
+// @import '{}/imports/ui/stylesheets/list-items.less';
+// @import '{}/imports/ui/stylesheets/message.less';
+// @import '{}/imports/ui/stylesheets/notification.less';
+//
+// // Individual components
+// @import "{}/imports/ui/components/lists-show.less";
+// @import "{}/imports/ui/components/loading.less";
+// @import "{}/imports/ui/pages/app-not-found.less";
+// @import "{}/imports/ui/accounts/accounts-templates.less";
