--- client/partials/nav.html
@@ -1,46 +1,0 @@
-<template name="nav">
-    <nav class="navbar navbar-default-whiteBG navbar-fixed-top">
-        <div class="container">
-            <!-- Brand and toggle get grouped for better mobile display -->
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-                    <span class="sr-only">Toggle navigation</span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                </button>
-                <!-- <a class="navbar-brand " href="/"> Junior Film Fleadh</a> -->
-                <a href="/" class="navbar-left logo-img" id="logo-img" ><img src="/img/logo.png"></a>
-            </div>
-
-            <!-- Collect the nav links, forms, and other content for toggling -->
-            <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-                <!-- <ul class="nav navbar-nav nav pull-right"> -->
-                <ul class="nav navbar-nav navbar-right">
-                    <li><a href="/viewFilms"><i class="fa fa-film" aria-hidden="true"></i> View Films</a></li>
-
-                    {{#if currentUser}}
-                        {{#if currentUser.submittedFilm}}
-                            <li><a href="/profilePage"><i class="fa fa-user" aria-hidden="true"></i> Profile</a></li>
-                        {{else}}
-                            <li><a href="/filmSubmission"><i class="fa fa-plus-circle" aria-hidden="true"></i> Submit</a></li>
-                        {{/if}}
-                        <li><a href="#" class="logout"><i class="fa fa-sign-out" aria-hidden="true"></i> Logout</a></li>
-                        <li><a href="{{#if currentUser.submittedFilm}}/profilePage{{else}}/filmSubmission{{/if}}">| {{currentUser.profile.user.name}} |</a></li>
-                    {{else}}
-                        <li><a href="/signUp_reg"><i class="fa fa-user" aria-hidden="true"></i> SignUp/Login</a></li>
-                    {{/if}}
-                </ul>
-
-            </div>
-        </div>
-    </nav>
-     <!-- <script>
-    $(document).ready(function() {
-        $('.navbar-collapse a').click(function() {
-            $(".navbar-collapse").collapse('hide');
-        });
-    });
-</script> -->
-
-</template>
