+++ imports/api/filmManagement/comments/server/publications.js
@@ -1,0 +1,2 @@
+console.log("DEBUG: ran publications.js");
+
