+++ imports/api/filmManagement/films/server/publications.js
@@ -1,0 +1,2 @@
+console.log("DEBUG: ran publications.js");
+
