+++ imports/startup/both/index.js
@@ -1,0 +1,3 @@
+// nothing here yet
+
+// NOTE: the both dir is used for AccountsTemplate.configure:  https://github.com/meteor/todos/blob/master/imports/startup/both/useraccounts-configuration.js
