+++ imports/startup/client/main.html
@@ -17,7 +17,7 @@
 </head>
 
 <body>
-    <!-- // TODO: move this into router's responsibility -->
+    <!-- // TODO: move this into router's responsibility. only <head> should be in this file -->
     {{> nav}}
