+++ imports/startup/client/main.js
@@ -1,0 +1,6 @@
+import { Template } from 'meteor/templating';
+import { ReactiveVar } from 'meteor/reactive-var';
+
+/*global _*/
+
+import './main.html';
