+++ imports/startup/server/register-apis.js
@@ -1,0 +1,3 @@
+console.log("DEBUG: ran register-apis.js");
+console.log("DEBUG: next log should be: filmManagement/index.js");
+import '../../api/filmManagement';
